adding calls

+++ src/main/java/ua/com/smiddle/cti/messages/model/messages/agent_events/calls/AgentPreCallEvent.java
@@ -132,7 +132,7 @@ public class AgentPreCallEvent extends Header {
                     .putInt(serviceID)
                     .putInt(skillGroupNumber)
                     .putInt(skillGroupID)
-                    .putInt(skillGroupPriority)
+                    .putShort(skillGroupPriority)
                     .putInt(mrdid);
             for (FloatingField field : floatingFields) {
                 field.serializeField(buffer);
